requirements.txt
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1673535674859/work
audioread==3.0.1
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
brotlipy==0.7.0
cachetools==5.3.1
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1700303426725/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
coloredlogs==15.0.1
contextlib2==21.6.0
cryptography @ file:///croot/cryptography_1677533068310/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
diffdist==0.1
einops==0.6.1
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
et-xmlfile==1.1.0
fairscale==0.4.6
filelock==3.12.2
flatbuffers==23.3.3
fonttools==4.25.0
fsspec==2023.1.0
ftfy==6.1.1
google-auth==2.22.0
google-auth-oauthlib==0.4.6
grpcio @ file:///tmp/build/80754af9/grpcio_1637578017926/work
h5py==3.8.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna @ file:///croot/idna_1666125576474/work
ImageHash==4.3.1
imageio==2.31.2
imageio-ffmpeg==0.4.9
imgaug==0.4.0
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1653252814274/work
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1620912939357/work/dist/ipykernel-5.5.5-py3-none-any.whl
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1651240553635/work
ipython-genutils==0.2.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
joblib @ file:///croot/joblib_1666298844297/work
jupyter-client==5.3.4
jupyter-core==4.5.0
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
kornia==0.6.12
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.39.1
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1679584000376/work
matplotlib @ file:///opt/conda/conda-bld/matplotlib-suite_1660167928326/work
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
ml-collections==0.1.1
moviepy==1.0.3
mpmath==1.2.1
msgpack==1.0.5
munkres==1.1.4
networkx==2.6.3
numba==0.56.4
numexpr @ file:///croot/numexpr_1668713893690/work
numpy==1.21.6
oauthlib==3.2.2
onnxruntime==1.14.1
open-clip-torch==2.24.0
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging @ file:///croot/packaging_1671697413597/work
pandas==1.3.5
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.4.0
platformdirs==4.0.0
ply==3.11
pooch==1.8.0
prefetch-generator==1.0.3
proglog==0.1.10
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1702399386289/work
protobuf==3.15.8
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1700607939962/work
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PyQt5-sip==12.11.0
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///croot/pytz_1671697431263/work
PyWavelets==1.3.0
PyYAML==6.0.1
pyzmq==19.0.2
regex==2023.12.25
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
safetensors==0.4.1
scikit-image==0.19.3
scikit-learn @ file:///tmp/build/80754af9/scikit-learn_1642601761909/work
scipy @ file:///opt/conda/conda-bld/scipy_1661390393401/work
seaborn==0.12.2
sentencepiece==0.1.99
shapely==2.0.2
sip @ file:///tmp/abs_44cd77b_pu/croots/recipe/sip_1659012365470/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0.post1
soundfile==0.12.1
soxr==0.3.7
submodlib==1.1.5
sympy==1.10.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tifffile==2021.11.2
timm==0.9.12
tokenizers==0.13.3
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
torch==1.13.1
torch-tb-profiler==0.4.1
torchaudio==0.13.1
torchlars==0.1.2
torchmetrics==0.11.4
torchsampler==0.1.2
torchvision==0.14.1
tornado @ file:///opt/conda/conda-bld/tornado_1662061693373/work
tqdm==4.65.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
transformers==4.30.2
typing_extensions==4.7.1
urllib3 @ file:///croot/urllib3_1673575502006/work
wcwidth==0.2.13
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1651670883478/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1677313463193/work
